use double not float
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@59989 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/gui/dialog/options/model/ObjectSettingsModel.java
@@ -473,8 +473,8 @@ abstract public class ObjectSettingsModel {
     /**
      * @return the current alpha value of the geoElement
      */
-    public float getAlpha() {
-        return geoElement != null ? (float) geoElement.getAlphaValue() : 1;
+	public double getAlpha() {
+		return geoElement != null ? geoElement.getAlphaValue() : 1;
     }
 
     /**
